+++ src/main/java/io/github/zero88/exceptions/ErrorCode.java
@@ -2,6 +2,8 @@ package io.github.zero88.exceptions;
 
 import java.io.Serializable;
 
+import com.fasterxml.jackson.annotation.JsonValue;
+
 import lombok.NonNull;
 
 public interface ErrorCode extends Serializable {
@@ -11,7 +13,10 @@ public interface ErrorCode extends Serializable {
     ErrorCode FILE_ERROR = new InternalErrorCode("FILE_ERROR");
     ErrorCode REFLECTION_ERROR = new InternalErrorCode("REFLECTION_ERROR");
     ErrorCode URL_ERROR = new InternalErrorCode("URL_ERROR");
+    ErrorCode UNSUPPORTED = new InternalErrorCode("UNSUPPORTED");
+    ErrorCode INVALID_ARGUMENT = new InternalErrorCode("INVALID_ARGUMENT");
 
+    @JsonValue
     @NonNull String code();
 
 }