+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
+## [v1.0.1](https://github.com/zero88/java-utils/tree/release/1.0.1) (2020-12-15)
+
+### Changed
+
+- Make `ErrorCodeException` as super interface of `Exception`. 
+- Make `ErrorCode` is JSON serialize-able by `Jackson` and able to compare with `String`  
+- Improve build
+
 ## [v1.0.0](https://github.com/zero88/java-utils/tree/release/1.0.0) (2020-09-17)